slackrtm: split plugin-specific exceptions into its own file
needed for slack command processor
