+++ hangupsbot/plugins/slackrtm/exceptions.py
@@ -1,0 +1,18 @@
+class ParseError(Exception):
+    pass
+
+
+class AlreadySyncingError(Exception):
+    pass
+
+
+class NotSyncingError(Exception):
+    pass
+
+
+class ConnectionFailedError(Exception):
+    pass
+
+
+class IncompleteLoginError(Exception):
+    pass
